bmap/requirements.txt
freetype-py
pillow
